refactor: correct import statements for client room entities
