+++ core/data/src/main/java/com/mifos/core/data/pagingSource/ClientChargesPagingSource.kt
@@ -11,9 +11,9 @@ package com.mifos.core.data.pagingSource
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
+import com.mifos.core.model.objects.clients.Page
 import com.mifos.core.network.datamanager.DataManagerCharge
 import com.mifos.room.entities.client.Charges
-import com.mifos.room.entities.client.Page
 import kotlinx.coroutines.suspendCancellableCoroutine
 import rx.Subscriber
 import rx.android.schedulers.AndroidSchedulers
