+++ core/data/src/main/java/com/mifos/core/data/repository/ClientIdentifierDialogRepository.kt
@@ -9,9 +9,9 @@
  */
 package com.mifos.core.data.repository
 
-import com.mifos.room.entities.noncore.IdentifierCreationResponse
-import com.mifos.room.entities.noncore.IdentifierPayload
-import com.mifos.room.entities.noncore.IdentifierTemplate
+import com.mifos.core.model.objects.noncoreobjects.IdentifierCreationResponse
+import com.mifos.core.model.objects.noncoreobjects.IdentifierPayload
+import com.mifos.core.model.objects.noncoreobjects.IdentifierTemplate
 
 /**
  * Created by Aditya Gupta on 16/08/23.
