+++ core/data/src/main/java/com/mifos/core/data/repositoryImp/CreateNewClientRepositoryImp.kt
@@ -10,14 +10,14 @@
 package com.mifos.core.data.repositoryImp
 
 import com.mifos.core.data.repository.CreateNewClientRepository
-import com.mifos.core.entity.client.Client
-import com.mifos.core.entity.client.ClientPayload
-import com.mifos.core.entity.organisation.Office
-import com.mifos.core.entity.organisation.Staff
-import com.mifos.core.entity.templates.clients.ClientsTemplate
 import com.mifos.core.network.datamanager.DataManagerClient
 import com.mifos.core.network.datamanager.DataManagerOffices
 import com.mifos.core.network.datamanager.DataManagerStaff
+import com.mifos.room.entities.client.Client
+import com.mifos.room.entities.client.ClientPayload
+import com.mifos.room.entities.organisation.Office
+import com.mifos.room.entities.organisation.Staff
+import com.mifos.room.entities.templates.clients.ClientsTemplate
 import kotlinx.coroutines.flow.Flow
 import okhttp3.MultipartBody
 import okhttp3.ResponseBody
