+++ core/database/src/main/java/com/mifos/room/db/MifosDatabase.kt
@@ -14,6 +14,7 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.mifos.room.dao.ColumnValueDao
 import com.mifos.room.dao.LoanDao
+import com.mifos.room.dao.ClientDao
 import com.mifos.room.entities.PaymentTypeOption
 import com.mifos.room.entities.accounts.loans.ActualDisbursementDate
 import com.mifos.room.entities.accounts.loans.LoanRepaymentRequest
@@ -21,6 +22,9 @@ import com.mifos.room.entities.accounts.loans.LoanWithAssociations
 import com.mifos.room.entities.accounts.loans.Status
 import com.mifos.room.entities.accounts.loans.Summary
 import com.mifos.room.entities.accounts.loans.Timeline
+import com.mifos.room.entities.client.ClientsAttendance
+import com.mifos.room.entities.client.ChargeTimeType
+import com.mifos.room.entities.client.ChargeCalculationType
 import com.mifos.room.entities.noncore.ColumnValue
 import com.mifos.room.entities.templates.loans.LoanRepaymentTemplate
 import com.mifos.room.utils.typeconverters.DueDateConverter
@@ -41,6 +45,11 @@ import com.mifos.room.utils.typeconverters.ServerTypesConverters
         Timeline::class,
         Status::class,
         Summary::class,
+        // client
+        ChargeCalculationType::class,
+        ChargeTimeType::class,
+        ClientsAttendance::class,
+        com.mifos.room.entities.client.Status::class,
     ],
     version = MifosDatabase.VERSION,
     exportSchema = true,
@@ -57,6 +66,7 @@ import com.mifos.room.utils.typeconverters.ServerTypesConverters
 abstract class MifosDatabase : RoomDatabase() {
     abstract fun columnValueDao(): ColumnValueDao
     abstract fun loanDao(): LoanDao
+    abstract fun clientDao(): ClientDao
 
     companion object {
         const val VERSION = 1
